moving the Game example to plutus-starter; minor cleanup
