+++ nix/pkgs/haskell/materialized-unix/.plan.nix/plutus-use-cases.nix
@@ -62,7 +62,6 @@
           "Plutus/Contracts/Currency"
           "Plutus/Contracts/Escrow"
           "Plutus/Contracts/Future"
-          "Plutus/Contracts/Game"
           "Plutus/Contracts/GameStateMachine"
           "Plutus/Contracts/ErrorHandling"
           "Plutus/Contracts/MultiSig"
@@ -126,9 +125,7 @@
             "Spec/ErrorHandling"
             "Spec/Escrow"
             "Spec/Future"
-            "Spec/Game"
             "Spec/GameStateMachine"
-            "Spec/Lib"
             "Spec/MultiSig"
             "Spec/MultiSigStateMachine"
             "Spec/PingPong"
@@ -184,9 +181,7 @@
             "Spec/ErrorHandling"
             "Spec/Escrow"
             "Spec/Future"
-            "Spec/Game"
             "Spec/GameStateMachine"
-            "Spec/Lib"
             "Spec/MultiSig"
             "Spec/MultiSigStateMachine"
             "Spec/PingPong"
