+++ plutus-use-cases/plutus-use-cases.cabal
@@ -32,7 +32,6 @@ library
         Plutus.Contracts.Currency
         Plutus.Contracts.Escrow
         Plutus.Contracts.Future
-        Plutus.Contracts.Game
         Plutus.Contracts.GameStateMachine
         Plutus.Contracts.ErrorHandling
         Plutus.Contracts.MultiSig
@@ -99,9 +98,7 @@ test-suite plutus-use-cases-test
         Spec.ErrorHandling
         Spec.Escrow
         Spec.Future
-        Spec.Game
         Spec.GameStateMachine
-        Spec.Lib
         Spec.MultiSig
         Spec.MultiSigStateMachine
         Spec.PingPong
@@ -165,9 +162,7 @@ executable plutus-use-cases-scripts
         Spec.ErrorHandling
         Spec.Escrow
         Spec.Future
-        Spec.Game
         Spec.GameStateMachine
-        Spec.Lib
         Spec.MultiSig
         Spec.MultiSigStateMachine
         Spec.PingPong
