+++ plutus-use-cases/test/Spec/MultiSigStateMachine.hs
@@ -14,8 +14,6 @@ import           Data.Foldable                         (traverse_)
 import           Test.Tasty                            (TestTree, testGroup)
 import qualified Test.Tasty.HUnit                      as HUnit
 
-import           Spec.Lib                              as Lib
-
 import qualified Ledger
 import qualified Ledger.Ada                            as Ada
 import qualified Ledger.Typed.Scripts                  as Scripts
@@ -54,8 +52,8 @@ tests =
         .&&. walletFundsChange w2 (Ada.lovelaceValueOf 10))
         (lockProposeSignPay 3 3)
 
-    , Lib.goldenPir "test/Spec/multisigStateMachine.pir" $$(PlutusTx.compile [|| MS.mkValidator ||])
-    , HUnit.testCase "script size is reasonable" (Lib.reasonable (Scripts.validatorScript $ MS.scriptInstance params) 51000)
+    , goldenPir "test/Spec/multisigStateMachine.pir" $$(PlutusTx.compile [|| MS.mkValidator ||])
+    , HUnit.testCase "script size is reasonable" (reasonable (Scripts.validatorScript $ MS.scriptInstance params) 51000)
     ]
 
 w1, w2, w3 :: EM.Wallet
