+++ plutus-use-cases/test/Spec/Rollup.hs
@@ -14,7 +14,7 @@ import           Ledger                        (pubKeyHash)
 import           Plutus.Contract.Trace
 
 import           Plutus.Contracts.Crowdfunding
-import           Plutus.Contracts.Game
+import qualified Spec.GameStateMachine
 import qualified Spec.Vesting
 
 import           Plutus.Trace.Emulator         (EmulatorTrace, runEmulatorStream)
@@ -32,9 +32,9 @@ tests = testGroup "showBlockchain"
           "test/Spec/renderCrowdfunding.txt"
           (render successfulCampaign)
      , goldenVsString
-          "renders a guess scenario sensibly"
+          "renders a game guess scenario sensibly"
           "test/Spec/renderGuess.txt"
-          (render guessTrace)
+          (render Spec.GameStateMachine.successTrace)
      , goldenVsString
           "renders a vesting scenario sensibly"
           "test/Spec/renderVesting.txt"
